requirements.txt
# Django
Django==3.2.6

# Database
#psycopg2===2.8.6
psycopg2-binary===2.8.6

# Envs
django-dotenv==1.4.2
s3-environ==0.0.5

# Rest Framework
djangorestframework==3.12.4
django-filter==2.4.0

# Routers
drf-nested-routers==0.93.3

# Database
dj_database_url==0.5.0

# Storages
django-storages==1.11.0

# Run Server
gunicorn==20.0.4
whitenoise==5.0.1

# CORS
django-cors-headers==3.7.0

# Rest Auth
django-rest-auth[with_social]==0.9.2

# Email
django-amazon-ses==4.0.0

# iPython
ipython==7.12.0
ipython-genutils==0.2.0

# Project Dependencies
requests==2.26.0
sentry-sdk==1.3.1